% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dalys}
\alias{dalys}
\title{Find DALYs with 95\% uncertainty intervals}
\usage{
dalys(
  incR,
  rate,
  pop,
  life,
  dw_A,
  dw_C,
  prog,
  dRate,
  duration,
  deaths,
  deathRate
)
}
\arguments{
\item{incR}{incident rates}

\item{rate}{used for incR eg. per 100 persons, per 100k}

\item{pop}{population-at-age, same age groups as incR}

\item{life}{life-expectancy-at-age, same age groups as incR}

\item{dw_A}{acute disability weights with uncertainty}

\item{dw_C}{chronic disability weights with uncertainty}

\item{prog}{probability of progression from acute to chronic}

\item{dRate}{discount rate as a percent}

\item{duration}{duration of disease}

\item{deaths}{death rates}

\item{deathRate}{rate used for deaths}
}
\value{
matrix
}
\description{
Find DALYs with 95\% uncertainty intervals
}
