% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.cellulitis2019}
\alias{data.cellulitis2019}
\title{Incident rates for cellulitis}
\format{
A data table containing 8532 observations of 5 variables.
\describe{
\item{location}{Region, subregion, country or area}
\item{age}{age groups, <1, 1 to 4, then 5-year age groups until 80 plus}
\item{val}{incident rate per 100,000 persons}
\item{upper}{upper bound of 95\% confidence interval}
\item{lower}{lower bound of 95\% confidence interval}
}
}
\source{
\url{http://ghdx.healthdata.org/gbd-results-tool}
}
\usage{
data.cellulitis2019
}
\description{
Incident rates (per 100,000 persons) for 2019 by country/region and age group.
Data set extracted from the Global Health Data Exchange (GHDx), created and
supported by the Institute for Health Metrics and Evaluation (IHME)
}
\keyword{datasets}
