% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.coveragebycountry}
\alias{data.coveragebycountry}
\title{Country-specific year of vaccine introduction and maximum coverage}
\format{
A data table containing 218 observations of 6 variables
\describe{
\item{Country}{Country}
\item{Region}{Region}
\item{Income.Group}{One of "Low income", "Lower middle income",
"Upper middle income" or "High income"}
\item{Year.of.Introduction}{Predicted year of vaccine introbution}
\item{Coverage}{As a proportion}
\item{Size.of Public.Market}{As a proportion}
}
}
\usage{
data.coveragebycountry
}
\description{
Shift health prediction of likely year of vaccine introduction for
each country and predicted maximum coverage proportion
}
\keyword{datasets}
