% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.life.fiji2019}
\alias{data.life.fiji2019}
\title{Probability of mortality - Fiji}
\format{
A data table containing 19 observations of 4 variables.
\describe{
\item{Age group}{age groups, <1, 1 to 4, then 5-year age groups until 85 plus}
\item{Both sexes}{combined probability}
\item{Male}{probability for males}
\item{Female}{probability for females}
}
}
\source{
\url{https://apps.who.int/gho/data/node.resources}
}
\usage{
data.life.fiji2019
}
\description{
Probability of dying between ages x and x+1 (2019). Data from the Global
Health Observatory, WHO
}
\keyword{datasets}
