% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.lifeEx.pred2100}
\alias{data.lifeEx.pred2100}
\title{Life expectancy - 2020 - 2100}
\format{
A data table containing 4016 observations of 25 variables
\describe{
\item{Location}{Region, subregion, country or area}
\item{Country code}{United Nations M49 numeric codes}
\item{Period}{Year range}
\item{0, 1, 5, ..., 95, 100+}{life expactancy at age}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Mortality/}
}
\usage{
data.lifeEx.pred2100
}
\description{
Life expectancy at exact age for both sexes combined by region and period.
Data set extracted from United Nations data "Life Expectancy at exact age x (ex) - Both Sexes"
obtained from source,
}
\keyword{datasets}
