% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mortality.1950_to_2020}
\alias{data.mortality.1950_to_2020}
\title{Probability of mortality - 1950 - 2020}
\format{
A data table containing 73794 observations of 5 variables.
\describe{
\item{Location}{Region, subregion, country or area}
\item{Period}{Year range}
\item{Age (x)}{Start of age range}
\item{Age interval (n)}{age interval length}
\item{Probability of dying q(x,n)}{probability of all-cause mortality between ages x and x+n}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Mortality/}
}
\usage{
data.mortality.1950_to_2020
}
\description{
Probability of dying between ages x and x+n by country and period.
Data set extracted from United Nations data "Abridged Life Table - Both Sexes"
obtained from source.
}
\keyword{datasets}
