% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mortality.pred2100}
\alias{data.mortality.pred2100}
\title{Probability of mortality - 2050 - 2100}
\format{
A data table containing 31626 observations of 5 variables.
\describe{
\item{Location}{Region, subregion, country or area}
\item{Period}{country or area}
\item{Sex}{female, male}
\item{Age group}{age groups, <1, 1 to 4, then 5-year age groups until 85 plus}
\item{Value}{probability of mortality}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Mortality/}
}
\usage{
data.mortality.pred2100
}
\description{
Probability of dying between ages x and x+n by country and period.
Data set extracted from United Nations data "Abridged Life Table - Both Sexes"
obtained from source.
}
\keyword{datasets}
