% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mortality2019}
\alias{data.mortality2019}
\title{Probability of mortality}
\format{
A data table containing 6954 observations of 5 variables.
\describe{
\item{Region}{Europe, Western Pacific, Americas, Eastern Mediterranean,
Africa and South-East Asia}
\item{Location}{country or area}
\item{Sex}{female, male}
\item{Age group}{age groups, <1, 1 to 4, then 5-year age groups until 85 plus}
\item{Value}{probability of mortality}
}
}
\source{
\url{https://www.who.int/data/gho/data/indicators/indicator-details/GHO/gho-ghe-life-tables-by-country}
}
\usage{
data.mortality2019
}
\description{
Probability of dying between ages x and x+1 (2019). Data from the Global
Health Observatory, WHO
}
\keyword{datasets}
