% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.popbyage.pred}
\alias{data.popbyage.pred}
\title{Population by age - 2020 - 2100}
\format{
A data table containing 20331 observations of 104 variables.
\describe{
\item{Country}{Region, subregion, country or area}
\item{Country code}{iso3 country code}
\item{Year}{2020 - 2100}
\item{0, 1, ..., 100}{population in 1-year age groups}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Interpolated/}
}
\usage{
data.popbyage.pred
}
\description{
Total interpolated population (both sexes combined) for 2020 - 2100
by 1-year age groups (in thousands).
Data set extracted from United Nations data "Annual Population by Age - Both Sexes"
obtained from source.
}
\keyword{datasets}
