% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.popbyage2020}
\alias{data.popbyage2020}
\title{2020 Population by age group}
\format{
A data table containing 251 observations of 22 variables.
\describe{
\item{Country}{Region, subregion, country or area}
\item{0-4, 5-9, ..., 95-99, 100+}{5-year age groups}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
data.popbyage2020
}
\description{
Total population (both sexes combined) by 5-year age group (in thousands).
Data set extracted from "Annual Population by Age Groups - Both Sexes" obtained
from source.
}
\keyword{datasets}
