% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.region}
\alias{data.region}
\title{Countries by United Nations regions
World Development Indicators and Other World Bank Data updated April 2021
From R package WDI}
\format{
A data table containing 304 observations of 4 variables.
\describe{
\item{Code}{iso3 country code}
\item{Country}{Country or area name}
\item{Region}{UN region name}
\item{IncomeGroup}
}
}
\usage{
data.region
}
\description{
Countries by United Nations regions
World Development Indicators and Other World Bank Data updated April 2021
From R package WDI
}
\keyword{datasets}
