% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findCalendarYear}
\alias{findCalendarYear}
\title{Restructure output to get in calendar year format including cohorts
born from 2020}
\usage{
findCalendarYear(modelResults2020, modelResultsFromVaccYear)
}
\arguments{
\item{model2020}{list of model results for cohorts from 2020}

\item{modelVaccYear}{list of model results for cohorts from
year of vaccine introduction}
}
\value{
list
}
\description{
Restructure output to get in calendar year format including cohorts
born from 2020
}
