% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findDalys}
\alias{findDalys}
\title{Calculate DALYs for both pre-vaccination and vaccination scenario. If
relevant, also return numbers of deaths for both scenarios}
\usage{
findDalys(
  condition,
  noVacc_counts,
  vacc_counts,
  daly_weights,
  location,
  yearV,
  pYears,
  vaccAge = 0,
  maxAge = 99,
  impType
)
}
\arguments{
\item{condition}{condition name}

\item{noVacc_counts}{number of cases of condition pre-vaccination}

\item{vacc_counts}{number of cases of condition after vaccination}

\item{daly_weights}{used in finding the YLD component of the DALYs it is the
"adjusted" disability weighting taking into account
duration of illness. Not used for RHD}

\item{location}{country that matches UN data}

\item{yearV}{year of vaccine introduction}

\item{pYears}{no. of years to project}

\item{vaccAge}{age of vaccination. Default 0}

\item{maxAge}{maximum age of cohort}

\item{impType}{type of impact analysis: Calendar year, Year of birth, Year of vaccination}
}
\value{
list(matrix, matrix, matrix, matrix)
}
\description{
Calculate DALYs for both pre-vaccination and vaccination scenario. If
relevant, also return numbers of deaths for both scenarios
}
