% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findDeaths}
\alias{findDeaths}
\title{Find number of deaths under simple scaling vaccine scenario
NOTE: Not currently used in app}
\usage{
findDeaths(noVaccDeaths, conditions, vaccAge = 0, vaccEff, vaccDur)
}
\arguments{
\item{noVaccDeaths}{age-specific death rates per 100,000 persons}

\item{conditions}{condition names}

\item{vaccAge}{age of vaccination. Default 0}

\item{vaccEff}{effectiveness of vaccine, as a percentage}

\item{vaccDur}{no. of years that vaccine is effective}
}
\value{
numeric
}
\description{
Find number of deaths under simple scaling vaccine scenario
NOTE: Not currently used in app
}
