% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findICER}
\alias{findICER}
\title{Find incremental cost-effectiveness ratio (ICER)}
\usage{
findICER(c0, cV, e0, eV, popS, VV)
}
\arguments{
\item{c0}{cost of disease with no vaccine}

\item{cV}{cost of disease with vaccine}

\item{e0}{DALYs lost without vaccine}

\item{eV}{DALYs lost with vaccine}

\item{popS}{population size}

\item{VV}{value of vaccine}
}
\value{
numeric
}
\description{
Find incremental cost-effectiveness ratio (ICER)
}
