% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findVacValue}
\alias{findVacValue}
\title{Find value of vaccine}
\usage{
findVacValue(c0, cV, e0, eV, popS, ICER)
}
\arguments{
\item{c0}{cost of disease with no vaccine}

\item{cV}{cost of disease with vaccine}

\item{e0}{DALYs lost without vaccine}

\item{eV}{DALYs lost with vaccine}

\item{popS}{population size}

\item{ICER}{incremental cost-effectiveness ratio (ICER)}
}
\value{
numeric
}
\description{
Find value of vaccine
}
