% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{formatCalendarY}
\alias{formatCalendarY}
\title{Restructure output to get in calendar year format.}
\usage{
formatCalendarY(noVacc_data, vacc_data)
}
\arguments{
\item{vacc_data}{matrix results matrix (counts, DALYs, deaths, pop, yll or yld)
for vaccination scenario from year of vaccination}

\item{no_vacc_data}{matrix results (counts, DALYs, deaths, pop, yll or yld)
for 2020 no vaccination scenario}
}
\value{
list(matrix, matrix)
}
\description{
Restructure output to get in calendar year format.
}
