% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getConditionData}
\alias{getConditionData}
\title{Retrieve location and condition specific incidence, DALYs and deaths data}
\usage{
getConditionData(location, condition, metric = NA, prop = 1)
}
\arguments{
\item{location}{country or region}

\item{condition}{Cellulitis, Rheumatic Heart Disease, Impetigo, Pharyngitis
or Invasive infection}

\item{metric}{Rate, Number or Percent. Only used when extracting Cellulitis
or Rheumatic Heart Disease data}

\item{prop}{Proportion of incident data that is attributable to GAS}
}
\value{
list(data.frame, data.frame, data.frame)
}
\description{
Note: Only Cellulitis and Rheumatic Heart Disease have country-specific data
}
