% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCountries}
\alias{getCountries}
\title{Retrieve all countries in a particular region}
\usage{
getCountries(region)
}
\arguments{
\item{region}{UN region}
}
\value{
vector
}
\description{
Retrieve all countries in a particular region
}
