% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMorData}
\alias{getMorData}
\title{Retrieve probability of all-cause mortality}
\usage{
getMorData(location)
}
\arguments{
\item{location}{country that matches UN data}
}
\value{
data.frame
}
\description{
Retrieve probability of all-cause mortality
}
