% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getRateData}
\alias{getRateData}
\title{Retrieve location and condition specific incidence and deaths data}
\usage{
getRateData(location, condition)
}
\arguments{
\item{location}{country that matches UN data}

\item{condition}{either RHD or cellulitis}
}
\value{
list(data.frame, data.frame)
}
\description{
Retrieve location and condition specific incidence and deaths data
}
