% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{groupResults}
\alias{groupResults}
\title{Present results in age groups}
\usage{
groupResults(age_data)
}
\arguments{
\item{age_data}{cases, DALYs, deaths results by yearly age}
}
\value{
data.frame
}
\description{
Present results in age groups
}
