% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heemodModel}
\alias{heemodModel}
\title{Set up and run heemod Markov model}
\usage{
heemodModel(probM, dalysM, dR, costM, Initpop, ageInit, cycleT)
}
\arguments{
\item{probM}{matrix of transition probabilities over model time period}

\item{dalysM}{matrix of DALYS for each age and condition}

\item{dR}{discount rate as a percentage}

\item{costM}{matrix of costs for each age and condition}

\item{Initpop}{initial population size}

\item{ageInit}{age when vaccine administered}

\item{cycleT}{age at end of model run}
}
\value{
heemod model
}
\description{
Set up and run heemod Markov model for a maximum of 10 disease states
}
