% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{incProb}
\alias{incProb}
\title{Find incidence probabilities for vaccination scenario}
\usage{
incProb(prI, maxAge, V_Eff, V_Age, V_Dur, impType, pYears)
}
\arguments{
\item{prI}{probability of incidence at age without vaccine}

\item{V_Eff}{vaccine effectiveness, reduce probability of incidence by this proportion}

\item{V_Age}{age when vaccine is administered}

\item{V_Dur}{length of time that vaccine is effective for}

\item{impType}{type of impact analysis: Calendar year, Year of birth, Year of vaccination}

\item{pYears}{number of years to project from year of vaccine introduction}
}
\value{
matrix
}
\description{
Find incidence probabilities for vaccination scenario
}
