% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{incRates}
\alias{incRates}
\title{Find incidence rates with 95\% uncertainty intervals}
\usage{
incRates(careEp, pop, propAttr, rate = 1e+05, years = 1, grps = 1)
}
\arguments{
\item{careEp}{raw care episodes-at-age}

\item{pop}{population-at-age that care episodes relate to}

\item{propAttr}{prop of episodes-at-age that are attributable to GAS for each disease}

\item{rate}{eg. per 100 persons, per 100k}

\item{years}{how many years does data cover}

\item{grps}{data frame used to combine and present results in user provided groups}
}
\value{
matrix
}
\description{
Find incidence rates with 95\% uncertainty intervals
}
