% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeBarPlot}
\alias{makeBarPlot}
\title{Create bar plot (with uncertainty values) of current (2019) data}
\usage{
makeBarPlot(plot_data, ylabel, colFill = "steelblue")
}
\arguments{
\item{plot_data}{data frame with age group, val
and upper and lower 95\% CI of one of Incidencem DALYs or deaths}

\item{ylabel}{y-axis label}

\item{colFill}{colour of bars}
}
\value{
ggplot
}
\description{
Incidence, DALYs or deaths per 100,000 persons
}
