% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makePlot}
\alias{makePlot}
\title{Create line plot comparison between vaccine and no vaccine scenarios}
\usage{
makePlot(noVacc_mod, vacc_mod, conditions)
}
\arguments{
\item{noVacc_mod}{heemod model for no vaccine scenario}

\item{vacc_mod}{heemod model for vaccine scenario}

\item{conditions}{conditions modeled}
}
\value{
ggplot
}
\description{
Create line plot comparison between vaccine and no vaccine scenarios
}
