% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makePlotAvert}
\alias{makePlotAvert}
\title{Create plots of counts, deaths or DALYs averted}
\usage{
makePlotAvert(
  noVacc_data,
  vacc_data,
  ylabel,
  vAge,
  maxAge = 99,
  vYear,
  impType,
  pYears
)
}
\arguments{
\item{noVacc_data}{counts, deaths OR DALYs for pre-vaccine scenario}

\item{vacc_data}{counts, deaths OR DALYs for vaccine scenario}

\item{ylabel}{label for y-axis}

\item{vAge}{vaccination age}

\item{maxAge}{maximum age of cohort}

\item{vYear}{year of vaccine introduction}

\item{impType}{type of impact analysis: Calendar year, Year of birth, Year of vaccination}

\item{pYears}{no. of years to project}
}
\value{
ggplot
}
\description{
Create plots of counts, deaths or DALYs averted
}
