% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeTable}
\alias{makeTable}
\title{Create table of comparison statistics}
\usage{
makeTable(noVacc_mod, vacc_mod, conditions, initPop = 1e+05, valueVac = -1)
}
\arguments{
\item{noVacc_mod}{heemod model for no vaccine scenario}

\item{vacc_mod}{heemod model for vaccine scenario}

\item{conditions}{conditions modeled}

\item{initPop}{initial population size. Default 100,000}

\item{valueVac}{value of vaccine}
}
\value{
data.frame
}
\description{
Create table of comparison statistics
}
