% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{markovModel}
\alias{markovModel}
\title{REDUNDANT FUNCTION, WILL REMOVE
Run Markov models for vaccine impact scenario}
\usage{
markovModel(
  age,
  inc,
  rate,
  costs,
  dalys,
  dalyRate,
  mortality,
  initPop,
  dRate,
  nyears,
  vaccAge,
  vaccEff,
  vaccDur,
  valueVac
)
}
\arguments{
\item{age}{age group names and year range}

\item{inc}{incidence rates for each condition}

\item{rate}{rate used for inc}

\item{costs}{costs per person}

\item{dalys}{DALYs for each condition}

\item{dalyRate}{rate used for DALYs}

\item{mortality}{all-cause mortality rate at age}

\item{initPop}{initial population size}

\item{dRate}{discount rate as a percentage}

\item{nyears}{no. of years to project}

\item{vaccAge}{age of vaccination}

\item{vaccEff}{effectiveness of vaccine, as a percentage}

\item{vaccDur}{no. of years that vaccine is effective}

\item{valueVac}{value of the vaccine}
}
\value{
table of comparison values
}
\description{
Run Markov models for vaccine impact scenario and produce table
of comparison metrics. This function calls transProb and heemodModel
}
