% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{missingData}
\alias{missingData}
\title{Check for missing age-specific data and add 0 row}
\usage{
missingData(dataF)
}
\arguments{
\item{dataF}{data frame from getRateData}
}
\value{
data.frame
}
\description{
Check for missing age-specific data and add 0 row
}
