% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preVaxxModel}
\alias{preVaxxModel}
\title{Find pre-vaccination incidents for a particular cohort using the population
size at birth. Function also returns the modeled population values at each age.}
\usage{
preVaxxModel(probM, InitPop, vAge = 0, maxAge = 99)
}
\arguments{
\item{probM}{matrix of transition probabilities over model time period}

\item{InitPop}{population size at age of vaccination for cohort of interest}

\item{vAge}{vaccination age}

\item{maxAge}{maximum age of cohort}
}
\value{
list(matrix, matrix, matrix)
}
\description{
Find pre-vaccination incidents for a particular cohort using the population
size at birth. Function also returns the modeled population values at each age.
}
