% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{probUncondition}
\alias{probUncondition}
\title{Change incidence probabilities to being unconditional}
\usage{
probUncondition(Pr)
}
\arguments{
\item{Pr}{vector/row of combined incid prob and all-cause mortality}
}
\value{
vector
}
\description{
Change incidence probabilities to being unconditional
}
