% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{propModel}
\alias{propModel}
\title{Heemod alternative}
\usage{
propModel(probM, daly_weights, InitPop, initialAge = 0, maxAge = 99)
}
\arguments{
\item{probM}{matrix of transition probabilities over model time period}

\item{maxAge}{maximum age of cohort}

\item{dalyWeights}{matrix of partial DALY (pre-count calculation) for each age and condition
Note: This will be NA for RHD as separate function used}

\item{Initpop}{initial population size}

\item{ageInit}{inital age}
}
\value{
list(matrix, matrix, matrix)
}
\description{
Heemod alternative
}
