% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runModel}
\alias{runModel}
\title{Main function to produce results for vaccine impact scenario}
\usage{
runModel(
  location,
  condition,
  inc,
  rate = 1e+05,
  mortality,
  maxAge = 99,
  yearV,
  vaccAge = 0,
  vaccEff = 100,
  vaccDur = 10,
  waning = FALSE,
  ramp = FALSE,
  impType,
  pYears = 10,
  initPop
)
}
\arguments{
\item{location}{country that matches UN data}

\item{condition}{condition name}

\item{inc}{incidence rates for each condition}

\item{rate}{rate used for inc. Default 100,000}

\item{mortality}{all-cause mortality rate at age}

\item{maxAge}{maximum age of cohort}

\item{yearV}{year of vaccine introduction}

\item{vaccAge}{age of vaccination. Default 0}

\item{vaccEff}{effectiveness of vaccine, as a percentage}

\item{vaccDur}{no. of years that vaccine is effective}

\item{waning}{assume linear waning immunity, TRUE or FALSE}

\item{ramp}{assume linear 10 year ramp up to maximum efficacy from year of introduction}

\item{impType}{type of impact analysis: Calendar year, Year of birth, Year of vaccination}

\item{pYears}{no. of years to project}

\item{initPop}{population at birth for cohorts of interest}
}
\value{
list
}
\description{
Main function to produce results for vaccine impact scenario
}
