% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transProb}
\alias{transProb}
\title{Find transition probabilities at each age from age of vaccination for each state}
\usage{
transProb(probD, probMort)
}
\arguments{
\item{probD}{probability of disease at age without vaccine}

\item{probMort}{all-cause mortality for cohort at age}
}
\value{
matrix
}
\description{
Find transition probabilities at each age from age of vaccination for each state
}
