% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dalysRHD}
\alias{dalysRHD}
\title{Find DALYs for RHD}
\usage{
dalysRHD(noVacc_counts, vacc_counts, lifeEx, HIC_flag)
}
\arguments{
\item{noVacc_counts}{pre-vaccine numbers of incidents}

\item{vacc_counts}{vaccine scenario numbers of incidents}
}
\value{
list(matrix, matrix)
}
\description{
Find DALYs for RHD
}
