% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.popbyage2020}
\alias{data.popbyage2020}
\title{Population by age - 2020}
\format{
A data table containing 251 observations of 102 variables.
\describe{
\item{Country}{Region, subregion, country or area}
\item{0, 1, ..., 100}{population in 1-year age groups}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Interpolated/}
}
\usage{
data.popbyage2020
}
\description{
Total population (both sexes combined) for 2020 by 1-year age
groups (in thousands).
Data set extracted from United Nations data "Annual Population by Age - Both Sexes"
obtained from source.
}
\keyword{datasets}
