% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.region}
\alias{data.region}
\title{Countries by United Nations regions}
\format{
A data table containing 230 observations of 2 variables.
\describe{
\item{Country}{Country or area name}
\item{Region}{UN region name}
}
}
\source{
\url{https://population.un.org/wpp/Download/Files/4_Metadata/WPP2019_F01_LOCATIONS.}
}
\usage{
data.region
}
\description{
Countries by United Nations regions
}
\keyword{datasets}
