% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.rhd2019}
\alias{data.rhd2019}
\title{Incidence and death rates for rheumatic heart disease (RHD)}
\format{
A data table containing 18040 observations of 6 variables.
\describe{
\item{measure}{Either Incidence or Deaths}
\item{location}{Region, subregion, country or area}
\item{age}{age groups, <1, 1 to 4, then 5-year age groups until 80 to 84}
\item{val}{incident rate per 100,000 persons}
\item{upper}{upper bound of 95\% confidence interval}
\item{lower}{lower bound of 95\% confidence interval}
}
}
\source{
\url{http://ghdx.healthdata.org/gbd-results-tool}
}
\usage{
data.rhd2019
}
\description{
Incidence and death rates (per 100,000 persons) for 2019 by country/region
and age group for 205 locations.Data set extracted from the Global Health
Data Exchange (GHDx), created and supported by the Institute for Health
Metrics and Evaluation (IHME)
}
\keyword{datasets}
