% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getConditionData}
\alias{getConditionData}
\title{Retrieve location and condition specific incidence, DALYs and deaths data}
\usage{
getConditionData(location, condition, metric)
}
\arguments{
\item{location}{country that matches UN data}

\item{condition}{either RHD or cellulitis}

\item{metric}{Rate, Number or Percent}
}
\value{
list(data.frame, data.frame, data.frame)
}
\description{
Retrieve location and condition specific incidence, DALYs and deaths data
}
