% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getInitPop}
\alias{getInitPop}
\title{Find population at birth for cohorts to be modeled}
\usage{
getInitPop(location, yearV, pYears, ageV, impType)
}
\arguments{
\item{location}{country that matches UN data}

\item{yearV}{year of vaccine introduction}

\item{pYears}{number of years to project from year of vaccine introduction}

\item{ageV}{vaccination age}

\item{impType}{the type of impact analysis, Calendar year, Year of birth or Year of Vaccination}
}
\value{
data.frame
}
\description{
Find population at birth for cohorts to be modeled
}
