% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPopData}
\alias{getPopData}
\title{Retrieve location population data}
\usage{
getPopData(location, yearV, pYears)
}
\arguments{
\item{location}{country that matches UN data}

\item{yearV}{year of vaccine introduction}

\item{pYears}{number of years to project from year of vaccine introduction}
}
\value{
data.frame
}
\description{
Follow a cohort from age 0 in year of vaccine introduction for
a period of 85 years, retrieving population numbers for each year of life
}
