% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPopData}
\alias{getPopData}
\title{Retrieve location population data}
\usage{
getPopData(location)
}
\arguments{
\item{location}{country that matches UN data}
}
\value{
data.frame
}
\description{
Retrieve location population data
}
