% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{immunity}
\alias{immunity}
\title{Find proportion of population immmune}
\usage{
immunity(waning, duration, efficacy)
}
\arguments{
\item{waning}{assume linear waning immunity, TRUE or FALSE}

\item{duration}{if linear waning, number of years taken to reach 50\% immunity,
else number of years of full immunity (after which immunity drops to 0)}

\item{efficacy}{the resulting efficacy proportion, efficacy x coverage}
}
\value{
vector
}
\description{
Find proportion of population immmune
}
