% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_functions.R
\name{makeTable}
\alias{makeTable}
\title{Create table of comparison statistics}
\usage{
makeTable(
  noVacc_counts,
  vacc_counts,
  noVacc_dalys,
  vacc_dalys,
  noVacc_deaths,
  vacc_deaths
)
}
\arguments{
\item{noVacc_counts}{numbers of cases for pre-vaccine scenario}

\item{vacc_counts}{numbers of cases for vaccine scenario}

\item{noVacc_dalys}{DALYs for pre-vaccine scenario}

\item{vacc_dalys}{DALYs for vaccine scenario}

\item{noVacc_deaths}{numbers of deaths for pre-vaccine scenario}

\item{vacc_deaths}{numbers of deaths for vaccine scenario}
}
\value{
data.frame
}
\description{
Create table of comparison statistics
}
