% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runModel}
\alias{runModel}
\title{Run Markov models for vaccine impact scenario}
\usage{
runModel(
  conditions,
  inc,
  rate = 1e+05,
  costs = 1,
  dalys = 1,
  dalyRate = 1e+05,
  mortality,
  initPop = 1e+05,
  dRate = 0,
  nyears = -1,
  vaccAge = 0,
  vaccEff = 100,
  vaccDur = 10
)
}
\arguments{
\item{conditions}{condition names}

\item{inc}{incident rates for each condition}

\item{rate}{rate used for inc. Default 100,000}

\item{costs}{costs per person for each condition and age group.
nAgeGroups x (nConditions + 1) matrix, col1 = ageGroup,
If set to 1, all age groups and conditions set to cost 1/person}

\item{dalys}{DALYs for each condition and age group.
nAgeGroups x (nConditions + 1) matrix, col1 = ageGroup,
If set to 1, all age groups and conditions set to DALYs 1}

\item{dalyRate}{rate used for DALYs. Default 100,000}

\item{mortality}{all-cause mortality rate at age}

\item{initPop}{initial population size. Default 100,000}

\item{dRate}{discount rate as a percentage. Default 0\%}

\item{nyears}{no. of years to project}

\item{vaccAge}{age of vaccination. Default 0}

\item{vaccEff}{effectiveness of vaccine, as a percentage}

\item{vaccDur}{no. of years that vaccine is effective}
}
\value{
list
}
\description{
Run Markov models for vaccine impact scenario and produce table
of comparison metrics. This function calls transProb and heemodModel
}
