% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runModel}
\alias{runModel}
\title{Run Markov models for vaccine impact scenario
NOTE: Cost, DALYS and discounting not currently used.
Only written for one condition though some programming started to allow for
multiple condition to be modeled at the same time}
\usage{
runModel(
  location,
  condition,
  inc,
  propA = 1,
  rate = 1e+05,
  mortality,
  dRate = 0,
  maxAge = 99,
  yearV,
  vaccAge = 0,
  vaccEff = 100,
  vaccDur = 10,
  waning = FALSE,
  ramp = FALSE,
  impType,
  pYears = 10,
  initPop
)
}
\arguments{
\item{location}{country that matches UN data}

\item{condition}{condition name}

\item{inc}{incidence rates for each condition}

\item{propA}{proportion of incidents attributable to Strep A. Default 1.}

\item{rate}{rate used for inc. Default 100,000}

\item{mortality}{all-cause mortality rate at age}

\item{dRate}{discount rate as a percentage. Default 0\%}

\item{maxAge}{maximum age of cohort}

\item{yearV}{year of vaccine introduction}

\item{vaccAge}{age of vaccination. Default 0}

\item{vaccEff}{effectiveness of vaccine, as a percentage}

\item{vaccDur}{no. of years that vaccine is effective}

\item{waning}{assume linear waning immunity, TRUE or FALSE}

\item{ramp}{assume linear 10 year ramp up to maximum efficacy from year of introduction}

\item{impType}{type of impact analysis: Calendar year, Year of birth, Year of vaccination}

\item{pYears}{no. of years to project}

\item{initPop}{used for modeling if metric is Number}
}
\value{
list
}
\description{
Run Markov models for vaccine impact scenario and produce table
of comparison metrics. This function calls transProb and heemodModel
}
