% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transProb}
\alias{transProb}
\title{Find transition probabilities for Markov model}
\usage{
transProb(prM, lifeTable, maxAge, V_Eff, V_Age, V_Dur, impType)
}
\arguments{
\item{prM}{probability of disease without vaccine}

\item{lifeTable}{all-cause mortality}

\item{maxAge}{number of years over which transition probs. calculated}

\item{V_Eff}{vaccine effectiveness, reduce prob. by this proportion}

\item{V_Age}{age when vaccine is administered}

\item{V_Dur}{length of time that vaccine is effective for}

\item{impType}{type of impact analysis: Calendar year, Year of birth, Year of vaccination}
}
\value{
matrix
}
\description{
Find transition probabilities for Markov model
}
